+++ mocks/01-basic-login/password-expired-after-identify.json
@@ -1,0 +1,199 @@
+{
+   "version":"1.0.0",
+   "stateHandle":"02QnGduopjlQFbOWvKlmNtolZigSXhX9iWt8480Pum",
+   "expiresAt":"2021-05-17T21:31:10.000Z",
+   "intent":"LOGIN",
+   "remediation":{
+      "type":"array",
+      "value":[
+         {
+            "rel":[
+               "create-form"
+            ],
+            "name":"reenroll-authenticator",
+            "relatesTo":[
+               "$.currentAuthenticator"
+            ],
+            "href":"https://foo.okta.com/idp/idx/challenge/answer",
+            "method":"POST",
+            "produces":"application/ion+json; okta-version=1.0.0",
+            "value":[
+               {
+                  "name":"credentials",
+                  "type":"object",
+                  "form":{
+                     "value":[
+                        {
+                           "name":"passcode",
+                           "label":"New password",
+                           "secret":true
+                        }
+                     ]
+                  },
+                  "required":true
+               },
+               {
+                  "name":"stateHandle",
+                  "required":true,
+                  "value":"02QnGduopjlQFbOWvKlmNtolZigSXhX9iWt8480Pum",
+                  "visible":false,
+                  "mutable":false
+               }
+            ],
+            "accepts":"application/json; okta-version=1.0.0"
+         }
+      ]
+   },
+   "currentAuthenticator":{
+      "type":"object",
+      "value":{
+         "type":"password",
+         "key":"okta_password",
+         "id":"autksbgegGSb3LW2j5d6",
+         "displayName":"Password",
+         "methods":[
+            {
+               "type":"password"
+            }
+         ],
+         "settings":{
+            "complexity":{
+               "minLength":8,
+               "minLowerCase":0,
+               "minUpperCase":0,
+               "minNumber":0,
+               "minSymbol":0,
+               "excludeUsername":true,
+               "excludeAttributes":[
+                  
+               ]
+            },
+            "age":{
+               "minAgeMinutes":0,
+               "historyCount":4
+            }
+         }
+      }
+   },
+   "authenticators":{
+      "type":"array",
+      "value":[
+         {
+            "type":"password",
+            "key":"okta_password",
+            "id":"autksbgegGSb3LW2j5d6",
+            "displayName":"Password",
+            "methods":[
+               {
+                  "type":"password"
+               }
+            ]
+         }
+      ]
+   },
+   "authenticatorEnrollments":{
+      "type":"array",
+      "value":[
+         {
+            "type":"email",
+            "key":"okta_email",
+            "id":"eaen2ex9xYe9UKlsr5d6",
+            "displayName":"Email",
+            "methods":[
+               {
+                  "type":"email"
+               }
+            ]
+         },
+         {
+            "type":"password",
+            "key":"okta_password",
+            "id":"lae1g3ciasfOrw7Bl5d6",
+            "displayName":"Password",
+            "methods":[
+               {
+                  "type":"password"
+               }
+            ]
+         },
+         {
+            "type":"phone",
+            "key":"phone_number",
+            "id":"paen2godbghmQ2Ats5d6",
+            "displayName":"Phone",
+            "methods":[
+               {
+                  "type":"sms"
+               },
+               {
+                  "type":"voice"
+               }
+            ]
+         }
+      ]
+   },
+   "recoveryAuthenticator":{
+      "type":"object",
+      "value":{
+         "type":"password",
+         "key":"okta_password",
+         "id":"autksbgegGSb3LW2j5d6",
+         "displayName":"Password",
+         "methods":[
+            {
+               "type":"password"
+            }
+         ],
+         "settings":{
+            "complexity":{
+               "minLength":8,
+               "minLowerCase":0,
+               "minUpperCase":0,
+               "minNumber":0,
+               "minSymbol":0,
+               "excludeUsername":true,
+               "excludeAttributes":[
+                  
+               ]
+            },
+            "age":{
+               "minAgeMinutes":0,
+               "historyCount":4
+            }
+         }
+      }
+   },
+   "user":{
+      "type":"object",
+      "value":{
+         "id":"00un2onoa5JhS00qz5d6"
+      }
+   },
+   "cancel":{
+      "rel":[
+         "create-form"
+      ],
+      "name":"cancel",
+      "href":"https://foo.okta.com/idp/idx/cancel",
+      "method":"POST",
+      "produces":"application/ion+json; okta-version=1.0.0",
+      "value":[
+         {
+            "name":"stateHandle",
+            "required":true,
+            "value":"02QnGduopjlQFbOWvKlmNtolZigSXhX9iWt8480Pum",
+            "visible":false,
+            "mutable":false
+         }
+      ],
+      "accepts":"application/json; okta-version=1.0.0"
+   },
+   "app":{
+      "type":"object",
+      "value":{
+         "name":"oidc_client",
+         "label":"Laura - My Web App",
+         "id":"foo"
+      }
+   }
+}
