+++ mocks/01-basic-login/two-steps-login/challenge-response.json
@@ -1,0 +1,73 @@
+{
+   "version":"1.0.0",
+   "stateHandle":"021ASG5DWgfcbJZ5jRU_KzF-K-6EHgE_Cq9o48TEZ4",
+   "expiresAt":"2021-05-17T16:12:37.000Z",
+   "intent":"LOGIN",
+   "user":{
+      "type":"object",
+      "value":{
+         "id":"foo"
+      }
+   },
+   "cancel":{
+      "rel":[
+         "create-form"
+      ],
+      "name":"cancel",
+      "href":"https://foo.okta.com/idp/idx/cancel",
+      "method":"POST",
+      "produces":"application/ion+json; okta-version=1.0.0",
+      "value":[
+         {
+            "name":"stateHandle",
+            "required":true,
+            "value":"021ASG5DWgfcbJZ5jRU_KzF-K-6EHgE_Cq9o48TEZ4",
+            "visible":false,
+            "mutable":false
+         }
+      ],
+      "accepts":"application/json; okta-version=1.0.0"
+   },
+   "app":{
+      "type":"object",
+      "value":{
+         "name":"oidc_client",
+         "label":"Laura - My Web App",
+         "id":"foo"
+      }
+   },
+   "successWithInteractionCode":{
+      "rel":[
+         "create-form"
+      ],
+      "name":"issue",
+      "href":"https://foo.okta.com/oauth2/foo/v1/token",
+      "method":"POST",
+      "value":[
+         {
+            "name":"grant_type",
+            "required":true,
+            "value":"interaction_code"
+         },
+         {
+            "name":"interaction_code",
+            "required":true,
+            "value":"u1bB0tA9bXIeHeESnDse_AvMrfEdSkfaSesemey14e4"
+         },
+         {
+            "name":"client_id",
+            "required":true,
+            "value":"foo"
+         },
+         {
+            "name":"client_secret",
+            "required":true
+         },
+         {
+            "name":"code_verifier",
+            "required":true
+         }
+      ],
+      "accepts":"application/x-www-form-urlencoded"
+   }
+}
