+++ mocks/01-basic-login/two-steps-login/identify-response.json
@@ -1,0 +1,145 @@
+{
+   "version":"1.0.0",
+   "stateHandle":"021ASG5DWgfcbJZ5jRU_KzF-K-6EHgE_Cq9o48TEZ4",
+   "expiresAt":"2021-05-17T16:16:35.000Z",
+   "intent":"LOGIN",
+   "remediation":{
+      "type":"array",
+      "value":[
+         {
+            "rel":[
+               "create-form"
+            ],
+            "name":"challenge-authenticator",
+            "relatesTo":[
+               "$.currentAuthenticatorEnrollment"
+            ],
+            "href":"https://foo.okta.com/idp/idx/challenge/answer",
+            "method":"POST",
+            "produces":"application/ion+json; okta-version=1.0.0",
+            "value":[
+               {
+                  "name":"credentials",
+                  "type":"object",
+                  "form":{
+                     "value":[
+                        {
+                           "name":"passcode",
+                           "label":"Password",
+                           "secret":true
+                        }
+                     ]
+                  },
+                  "required":true
+               },
+               {
+                  "name":"stateHandle",
+                  "required":true,
+                  "value":"021ASG5DWgfcbJZ5jRU_KzF-K-6EHgE_Cq9o48TEZ4",
+                  "visible":false,
+                  "mutable":false
+               }
+            ],
+            "accepts":"application/json; okta-version=1.0.0"
+         }
+      ]
+   },
+   "currentAuthenticatorEnrollment":{
+      "type":"object",
+      "value":{
+         "recover":{
+            "rel":[
+               "create-form"
+            ],
+            "name":"recover",
+            "href":"https://foo.okta.com/idp/idx/recover",
+            "method":"POST",
+            "produces":"application/ion+json; okta-version=1.0.0",
+            "value":[
+               {
+                  "name":"stateHandle",
+                  "required":true,
+                  "value":"021ASG5DWgfcbJZ5jRU_KzF-K-6EHgE_Cq9o48TEZ4",
+                  "visible":false,
+                  "mutable":false
+               }
+            ],
+            "accepts":"application/json; okta-version=1.0.0"
+         },
+         "type":"password",
+         "key":"okta_password",
+         "id":"lae1g3ciasfOrw7Bl5d6",
+         "displayName":"Password",
+         "methods":[
+            {
+               "type":"password"
+            }
+         ]
+      }
+   },
+   "authenticators":{
+      "type":"array",
+      "value":[
+         {
+            "type":"password",
+            "key":"okta_password",
+            "id":"autksbgegGSb3LW2j5d6",
+            "displayName":"Password",
+            "methods":[
+               {
+                  "type":"password"
+               }
+            ]
+         }
+      ]
+   },
+   "authenticatorEnrollments":{
+      "type":"array",
+      "value":[
+         {
+            "type":"password",
+            "key":"okta_password",
+            "id":"lae1g3ciasfOrw7Bl5d6",
+            "displayName":"Password",
+            "methods":[
+               {
+                  "type":"password"
+               }
+            ]
+         }
+      ]
+   },
+   "user":{
+      "type":"object",
+      "value":{
+         "id":"foo"
+      }
+   },
+   "cancel":{
+      "rel":[
+         "create-form"
+      ],
+      "name":"cancel",
+      "href":"https://foo.okta.com/idp/idx/cancel",
+      "method":"POST",
+      "produces":"application/ion+json; okta-version=1.0.0",
+      "value":[
+         {
+            "name":"stateHandle",
+            "required":true,
+            "value":"021ASG5DWgfcbJZ5jRU_KzF-K-6EHgE_Cq9o48TEZ4",
+            "visible":false,
+            "mutable":false
+         }
+      ],
+      "accepts":"application/json; okta-version=1.0.0"
+   },
+   "app":{
+      "type":"object",
+      "value":{
+         "name":"oidc_client",
+         "label":"Laura - My Web App",
+         "id":"foo"
+      }
+   }
+}
