+++ mocks/common/interact-response.json
@@ -1,0 +1,1 @@
+{"interaction_handle":"AcSDZw_kwcQtFUwLAgyUyBzl_OifS5Nn6IZQ_X1WXWI"}
