+++ src/Okta.Idx.Sdk.UnitTests/MockResponse.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Okta.Idx.Sdk.UnitTests
+{
+    public class MockResponse
+    {
+        public int StatusCode { get; set; }
+
+        public string Response { get; set; }
+    }
+}
