+++ src/direct-auth-idx/direct-auth-idx/packages.config
@@ -32,7 +32,7 @@
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net48" />
   <package id="Modernizr" version="2.8.3" targetFramework="net48" />
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net48" />
-  <package id="Okta.Sdk.Abstractions" version="3.0.2" targetFramework="net48" />
+  <package id="Okta.Sdk.Abstractions" version="3.0.3" targetFramework="net48" />
   <package id="Owin" version="1.0" targetFramework="net48" />
   <package id="System.Buffers" version="4.4.0" targetFramework="net48" />
   <package id="System.Collections.NonGeneric" version="4.3.0" targetFramework="net48" />
